### 2017-05-29 #269 @whedon
[PRE REVIEW]: RefManageR: Import and Manage BibTeX and BibLaTeX References in R
**Submitting author:** @mwmclean (<a href="http://orcid.org/0000-0002-7891-9645">Mathew McLean</a>)
**Repository:** <a href="https://github.com/mwmclean/RefManageR" target ="_blank">https://github.com/mwmclean/RefManageR</a>
**Version:** v0.13.4
