### 2017-05-29 #269 @arfon
👋 @noamross - would this package be suitable for rOpenSci?
