### 2017-05-31 #269 @noamross
Yes, this may be eligible for rOpenSci under our reproducibility and data access categories.  @mwmclean, if you are interested in joint submission, you may find more information [here](https://github.com/ropensci/onboarding#how-to-submit-your-package-for-review).
