### 2017-06-01 a94cdfe
CrossRef funs switched to jsonlite,httr; old API not working
