+++ R/GetDOIs.R
@@ -4,8 +4,9 @@
 #' \code{BibEntry} object.
 #' @param bib a \code{BibEntry} object
 #' @return \code{bib} with any found DOI's added in the \sQuote{doi} field
-#' @importFrom RJSONIO toJSON fromJSON
-#' @importFrom RCurl postForm
+#' @importFrom jsonlite toJSON fromJSON
+## @importFrom RCurl postForm
+#' @importFrom httr POST content http_error
 #' @keywords database
 #' @export
 #' @seealso \code{\link{ReadCrossRef}}
@@ -14,7 +15,7 @@
 #' \sQuote{doi} field will be searched for.
 #' @references \url{http://search.crossref.org/help/api}
 #' @examples
-#' if (interactive() && url.exists("http://search.crossref.org")){
+#' if (interactive() && !http_error("http://search.crossref.org")){
 #'   BibOptions(check.entries = FALSE, sorting = "none")
 #'   bib <- ReadBib(system.file("Bib", "RJC.bib", package = "RefManageR"))[1:5]
 #'   bib <- GetDOIs(bib)
@@ -30,20 +31,23 @@ GetDOIs <- function(bib){
     json.bib <- toJSON(FormatEntryForCrossRef(bib[[missing.dois.pos]]))
     headers <- list('Accept' = 'application/json', 'Content-Type' = 'application/json')
 
-    json.res <- postForm("http://search.crossref.org/links",
-                      .opts = list(postfields = json.bib, httpheader = headers))
-    json.res <- try(fromJSON(json.res), TRUE)
-    if (inherits(json.res, "try-error") || json.res[[2]] == FALSE)
-      message("Error with query.")
+    ## json.res <- postForm("http://search.crossref.org/links",
+    ##                   .opts = list(postfields = json.bib, httpheader = headers))
+    ## json.res <- try(fromJSON(json.res), TRUE)
+    json.res <- httr::POST("http://search.crossref.org/links", body = json.bib,
+                           config = list(add_headers = headers), encode = "json")
+    json.res <- try(content(json.res, type = "application/json", encoding = "UTF-8"), TRUE)
+    if (inherits(json.res, "try-error") || !json.res[[2L]])
+      message("Failed to retrieve any DOIs.")
     else{
-      matches <- sapply(json.res[[1]], "[[", "match")
+      matches <- sapply(json.res[[1L]], "[[", "match")
       if (!any(matches))
         message("No matches.")
       else{
         message(paste0("Matches for entries at positions ",
                        paste0(missing.dois.pos[matches], collapse = ", "), "."))
         bib$doi[missing.dois.pos[matches]] <- sub("http://dx.doi.org/", "",
-                                                   sapply(json.res[[1]], "[[",
+                                                   sapply(json.res[[1L]], "[[",
                                                           "doi")[matches], useBytes = TRUE)
       }
     }
