+++ R/ReadCrossRef.R
@@ -53,8 +53,9 @@
 #' \code{"award.funder"}, \code{"assertion-group"}, \code{"assertion"}, \code{"affiliation"},
 #' \code{"has-affiliation"}, \code{"alternative-id"}, and \code{"article-number"}.  See the first
 #' reference for a description of their meanings.
-#' @importFrom RJSONIO fromJSON
-#' @importFrom RCurl getForm getURLContent url.exists curlUnescape
+#' @importFrom jsonlite fromJSON
+#' @importFrom httr GET content http_error add_headers
+#' @importFrom utils URLdecode
 #' @export
 #' @keywords database
 #' @seealso \code{\link{ReadZotero}}, \code{\link{BibEntry}}, \code{\link{GetDOIs}},
@@ -63,7 +64,7 @@
 #' @references Newer API: \url{https://github.com/CrossRef/rest-api-doc/blob/master/rest_api.md},
 #' Older API: \url{http://search.crossref.org/help/api}
 #' @examples
-#' if (interactive() && url.exists("http://search.crossref.org/")){
+#' if (interactive() && !http_error("http://search.crossref.org/")){
 #'   BibOptions(check.entries = FALSE)
 #'   ## 3 results from the American Statistical Association involving "regression"
 #'   ReadCrossRef("regression", filter = list(prefix="10.1198"), limit = 3)
@@ -98,8 +99,9 @@ ReadCrossRef <- function(query = "", filter = list(), limit = 5, offset = 0,
     if (use.old.api){
       if (.is_not_nonempty_text(query))
           stop(gettextf("specify a valid %s", sQuote("query")))
-      results <- try(getForm("http://search.crossref.org/dois", q=query, year=year,
-                           sort=sort, rows=limit))        
+      headers <- list('Accept' = 'application/json', 'Content-Type' = 'application/json')
+      results <- try(GET("http://search.labs.crossref.org/dois", q=query, year=year,
+                           sort=sort, rows=limit, config = list(add_headers = headers)), TRUE)        
     }else{
       params <- list(rows = limit, sort = sort, offset = offset)
       if (!.is_not_nonempty_text(query))
@@ -110,12 +112,12 @@ ReadCrossRef <- function(query = "", filter = list(), limit = 5, offset = 0,
 
       if (length(filter))
           params$filter <- paste(paste0(names(filter),":",filter), collapse = ",")          
-      results <- try(getForm("http://api.crossref.org/works", .params=params))
+      results <- try(GET("http://api.crossref.org/works", query=params))
     }
     if (inherits(results, "try-error"))
-      stop(gettextf("RCurl failed to GET results from CrossRef: %s", geterrmessage()))
+      stop(gettextf("httr failed to GET results from CrossRef: %s", geterrmessage()))
 
-      fromj <- RJSONIO::fromJSON(results)
+    fromj <- content(results, type = "application/json", encoding = "UTF-8")
     if (!use.old.api)
         fromj <- fromj$message$items
     num.res <- min(limit, length(fromj))
@@ -127,7 +129,7 @@ ReadCrossRef <- function(query = "", filter = list(), limit = 5, offset = 0,
     }
 
     if (num.res > 0L){
-      if (!use.old.api && is.na(.doi)){
+        if (!use.old.api && is.na(.doi)){
           res <- lapply(fromj, ParseCrossRef)
           good <- which(sapply(res, function(e){
               good <- e$score >= min.relevance
@@ -179,7 +181,7 @@ ReadCrossRef <- function(query = "", filter = list(), limit = 5, offset = 0,
[…]
