+++ NAMESPACE
@@ -53,14 +53,9 @@ import(bibtex)
 import(methods)
 importFrom(RCurl,curlEscape)
 importFrom(RCurl,curlOptions)
-importFrom(RCurl,curlUnescape)
 importFrom(RCurl,getForm)
 importFrom(RCurl,getURL)
-importFrom(RCurl,getURLContent)
 importFrom(RCurl,postForm)
-importFrom(RCurl,url.exists)
-importFrom(RJSONIO,fromJSON)
-importFrom(RJSONIO,toJSON)
 importFrom(XML,free)
 importFrom(XML,getNodeSet)
 importFrom(XML,htmlParse)
@@ -70,6 +65,12 @@ importFrom(XML,xmlParse)
 importFrom(XML,xmlValue)
 importFrom(XML,xpathApply)
 importFrom(httr,GET)
+importFrom(httr,POST)
+importFrom(httr,add_headers)
+importFrom(httr,content)
+importFrom(httr,http_error)
+importFrom(jsonlite,fromJSON)
+importFrom(jsonlite,toJSON)
 importFrom(lubridate,"%within%")
 importFrom(lubridate,int_end)
 importFrom(lubridate,int_start)
@@ -96,6 +97,7 @@ importFrom(tools,latexToUtf8)
 importFrom(tools,loadPkgRdMacros)
 importFrom(tools,parseLatex)
 importFrom(tools,toRd)
+importFrom(utils,URLdecode)
 importFrom(utils,as.person)
 importFrom(utils,bibentry)
 importFrom(utils,browseURL)
