### 2017-06-01 c06109d
old CrossRef API working with httr
