### 2017-06-01 540f6c9
ReadGS converted to xml2/httr
