+++ R/04InternalFunctions.R
@@ -625,21 +625,90 @@ ProcessArxiv <- function(arxinfo){
 }
 
 #' @keywords internal
-#' @importFrom XML xmlValue
 #' @importFrom stringr str_sub str_trim
 ParseGSCites2 <- function(l, encoding, check.entries=.BibOptions$check.entries){
   if (!length(l))
     return(list())
   td <- l[[1L]]
 
-  title <- xmlValue(td[[1L]], encoding)
-  author <- xmlValue(td[[2L]], encoding)
-  cited_by <- as.numeric(xmlValue(l[[2L]][[1L]], encoding))
+  title <- td[[1L]][[1L]]  # xmlValue(td[[1L]], encoding)
+  author <- td[[2L]][[1L]]  # xmlValue(td[[2L]], encoding)
+  cited_by <- as.numeric(l[[2L]][[1L]][[1]])
   if (is.na(cited_by))  # no citation yet
     cited_by <- "0"
-  src <- xmlValue(td[[3L]], encoding)
+  src <- td[[3L]][[1L]]  # xmlValue(td[[3L]], encoding)
 
-  year <- as.numeric(regmatches(src, regexpr("([12][0-9]{3}$)", src, useBytes = TRUE)))
+  year <- as.numeric(regmatches(td[[3L]][[2L]][[1]], regexpr("([12][0-9]{3}$)", src, useBytes = TRUE)))
+  first_digit <- as.numeric(regexpr("[\\[\\(]?\\d",
+                                    src, useBytes = TRUE)) - 1L
+  ids <- which(first_digit < 0L)
+  first_digit <- replace(first_digit, ids, str_length(src)[ids])
+  journal <- str_trim(str_sub(src, 1L, first_digit))
+  trailing_commas <- as.numeric(regexpr(",$", journal, useBytes = TRUE)) - 1L
+  ids <- which(trailing_commas < 0L)
+  trailing_commas <- replace(trailing_commas, ids,
+                             str_length(journal)[ids])
+  journal <- str_sub(journal, 1L, trailing_commas)
+  numbers <- str_trim(str_sub(src, first_digit + 1L,
+                              str_length(src)))
+  # handle '...' in title, journal, or authors
+  if (!identical(check.entries, FALSE)){
+    if (is.null(title <- CheckGSDots(title, title, check.entries)) ||
+          is.null(author <- CheckGSDots(author, title, check.entries)) ||
+          is.null(journal <- CheckGSDots(journal, title, check.entries)))
+      return(NA)
+  }
+
+  res <- list(title = title, author = author, cites = cited_by,
+              year = year)
+  if (!is.na(eprint <- regmatches(src, regexec("arXiv:([0-9.]*)", src,
+                                               useBytes = TRUE))[[1]][2])){
+    res$eprinttype <- "arxiv"
+    res$eprint <- eprint
+    res$url <- paste0("http://arxiv.org/abs/", eprint)
+    attr(res, "entry") <- "misc"
+  }else{
+    if (is.na(numbers) || numbers == "" || as.character(year) == numbers){
+      if (as.numeric(cited_by) < 10L){
+        attr(res, "entry") <- "report"
+        res$institution <- journal
+        res$type <- "techreport"
+      }else{
+        attr(res, "entry") <- "book"
+        res$publisher <- res$journal
+      }
+    }else{
+      res$journal <- journal
+      res$number <- numbers
+      attr(res, 'entry') <- 'article'
+      numbers <- ProcessGSNumbers(res$number)
+      res$number <- numbers$number
+      res$pages <- numbers$pages
+      res$volume <- numbers$volume
[…]
