+++ NAMESPACE
@@ -58,7 +58,6 @@ importFrom(RCurl,getURL)
 importFrom(RCurl,postForm)
 importFrom(XML,free)
 importFrom(XML,getNodeSet)
-importFrom(XML,htmlParse)
 importFrom(XML,xmlDoc)
 importFrom(XML,xmlGetAttr)
 importFrom(XML,xmlParse)
@@ -106,3 +105,5 @@ importFrom(utils,head)
 importFrom(utils,person)
 importFrom(utils,tail)
 importFrom(utils,toBibtex)
+importFrom(xml2,xml_find_all)
+importFrom(xml2,xml_text)
