+++ DESCRIPTION
@@ -21,6 +21,7 @@ Description: Provides tools for importing and working with bibliographic
 License: GPL-2 | GPL-3 | BSD_3_clause + file LICENSE
 Imports:
     XML,
+    xml2,
     RCurl,
     jsonlite,
     utils,
