### 2017-06-01 d131ba1
switch url.exists to http_error
