+++ R/ReadPubMed.R
@@ -37,7 +37,7 @@
 #' @references \url{https://www.ncbi.nlm.nih.gov/books/NBK25499/#chapter4.ESearch}
 #' @family pubmed
 #' @examples
-#' if (interactive() && url.exists("https://eutils.ncbi.nlm.nih.gov/"))
+#' if (interactive() && !http_error("https://eutils.ncbi.nlm.nih.gov/"))
 #'   ReadPubMed(query = "raymond carroll measurement error", retmax = 5, mindate = 1990)
 ReadPubMed <- function(query, database = 'PubMed', ...){
   .params <- list(...)
@@ -95,7 +95,7 @@ ReadPubMed <- function(query, database = 'PubMed', ...){
 #' @references \url{https://www.ncbi.nlm.nih.gov/books/NBK25500/}
 #' @family pubmed
 #' @examples
-#' if (interactive() && url.exists("https://eutils.ncbi.nlm.nih.gov/"))
+#' if (interactive() && !http_error("https://eutils.ncbi.nlm.nih.gov/"))
 #'   GetPubMedByID(c("11209037", "21245076"))
 GetPubMedByID <- function(id, db = 'pubmed', ...){
 
@@ -160,7 +160,7 @@ GetPubMedByID <- function(id, db = 'pubmed', ...){
 #' @keywords database
 #' @export
 #' @examples
-#' if (interactive() && url.exists("https://eutils.ncbi.nlm.nih.gov/")){
+#' if (interactive() && !http_error("https://eutils.ncbi.nlm.nih.gov/")){
 #'   file.name <- system.file("Bib", "RJC.bib", package="RefManageR")
 #'   bib <- ReadBib(file.name)
 #'   bib <- LookupPubMedID(bib[[101:102]])
@@ -450,7 +450,7 @@ ProcessPubMedBookResult <- function(article){
 #' @keywords database
 #' @export
 #' @examples
-#' if (interactive() && url.exists("https://eutils.ncbi.nlm.nih.gov/")){
+#' if (interactive() && !http_error("https://eutils.ncbi.nlm.nih.gov/")){
 #'   file.name <- system.file("Bib", "RJC.bib", package="RefManageR")
 #'   bib <- ReadBib(file.name)
 #'   LookupPubMedID(bib[[101:102]])
