+++ R/GetBibEntryWithDOI.R
@@ -11,9 +11,10 @@
 #' The bibliographic information returned by the search of the \url{http://dx.doi.org} API is temporarily
 #' written to a file and then read back into \code{R} and return as a \code{BibEntry} object.
 #' @references \url{http://www.doi.org/tools.html}
+#' @importFrom httr http_error
 #' @seealso \code{\link{ReadCrossRef}}, \code{\link{BibEntry}}
 #' @examples
-#' if (interactive() && url.exists("http://dx.doi.org/"))
+#' if (interactive() && !http_error("http://dx.doi.org/"))
 #'   GetBibEntryWithDOI(c("10.1016/j.iheduc.2003.11.004", "10.3998/3336451.0004.203"))
 GetBibEntryWithDOI <- function(doi, temp.file=tempfile(fileext = '.bib'), delete.file = TRUE){
   file.create(temp.file)
