### 2017-06-01 9f40b5b
remove RCurl from GetBibEntryWithDOI
