### 2017-06-01 d74db24
remove url.exists from Rd files
