+++ man/LookupPubMedID.Rd
@@ -27,7 +27,7 @@ For each entry a citation string is created using the fields journaltitle/journa
   \dQuote{pubmed} and the eprint field is assigned the ID.
 }
 \examples{
-if (interactive() && url.exists("https://eutils.ncbi.nlm.nih.gov/")){
+if (interactive() && !http_error("https://eutils.ncbi.nlm.nih.gov/")){
   file.name <- system.file("Bib", "RJC.bib", package="RefManageR")
   bib <- ReadBib(file.name)
   LookupPubMedID(bib[[101:102]])
