### 2017-06-01 cdc36d1
fix check for bad doi
