### 2017-06-01 d0c6de3
use misc type if no journal info in ReadGS
