### 2017-06-01 0ad79c4
rm deprecated testthat fun use
