### 2017-06-01 e7f73c5
fix truncated returned results at limit
