### 2017-06-01 121702e
fix case of only one result; don't simplify mapply
