+++ R/ReadPubMed.R
@@ -36,14 +36,16 @@
 #' The language of the entry will be returned in the field \dQuote{language} and the abstract will be returned in the field \dQuote{abstract}, if they are available.
 #' @references \url{https://www.ncbi.nlm.nih.gov/books/NBK25499/#chapter4.ESearch}
 #' @family pubmed
+#' @importFrom httr GET read_xml
+#' @importFrom xml2 xml_find_all xml_text
 #' @examples
 #' if (interactive() && !http_error("https://eutils.ncbi.nlm.nih.gov/"))
 #'   ReadPubMed(query = "raymond carroll measurement error", retmax = 5, mindate = 1990)
-ReadPubMed <- function(query, database = 'PubMed', ...){
+ReadPubMed <- function(query, database = "PubMed", ...){
   .params <- list(...)
-  bad.ind <- which(!names(.params) %in% c('usehistory', 'WebEnv', 'query_key',
-                                          'retstart', 'retmax', 'field',
-                                          'datetype', 'reldate', 'mindate', 'maxdate'))
+  bad.ind <- which(!names(.params) %in% c("usehistory", "WebEnv", "query_key",
+                                          "retstart", "retmax", "field",
+                                          "datetype", "reldate", "mindate", "maxdate"))
   .parms <- .params
   if (length(bad.ind)){
     warning("Invalid .params specified and will be ignored")
@@ -55,18 +57,20 @@ ReadPubMed <- function(query, database = 'PubMed', ...){
 
   .parms$term <- query
   .parms$db <- database
-  .parms$usehistory = 'y'
-  results <- try(getForm(base.url, .params = .parms))
-  if (inherits(results, 'try-error'))
-    return(NA)
+  .parms$usehistory = "y"
+  ## results <- try(getForm(base.url, .params = .parms))
+  results <- try(GET(base.url, query = .parms), TRUE)
 
-  tdoc <- xmlParse(results)
+  if (inherits(results, "try-error"))
+    stop("Unable to GET results")
 
-  webenv <- unlist(xpathApply(tdoc, '/eSearchResult/WebEnv', xmlValue))
-  query.key <- unlist(xpathApply(tdoc, '/eSearchResult/QueryKey', xmlValue))
-  ids <- unlist(xpathApply(tdoc, '/eSearchResult/IdList/Id', xmlValue))
+  tdoc <- read_xml(results, encoding = "UTF-8")
+
+  webenv <- xml_text(xml_find_all(tdoc, "//eSearchResult/WebEnv"))
+  query.key <- xml_text(xml_find_all(tdoc, "//eSearchResult/QueryKey"))
+  ids <- xml_text(xml_find_all(tdoc, "/eSearchResult/IdList/Id"))
   if (!length(ids)){
-    message('No Results.')
+    message("No Results.")
     return()
   }
 
@@ -88,8 +92,8 @@ ReadPubMed <- function(query, database = 'PubMed', ...){
 #' unless a collection title is present -- in which case the \code{bibtype} will be
 #' \dQuote{InBook} -- or there is no journal information returned for an article -- in
 #' which case the \code{bibtype} will be \dQuote{Misc}.
-#' @importFrom RCurl postForm
-#' @importFrom XML xmlParse getNodeSet
+#' @importFrom httr POST read_xml
+#' @importFrom xml2 xml_text xml_find_all
 #' @keywords database
 #' @export
 #' @references \url{https://www.ncbi.nlm.nih.gov/books/NBK25500/}
@@ -97,25 +101,27 @@ ReadPubMed <- function(query, database = 'PubMed', ...){
 #' @examples
[…]
