### 2017-06-01 2216dbf
rm use of S3method tag in roxygen
