+++ R/format.R
@@ -7,7 +7,8 @@
 # @return character vector containing formatted BibEntry object.
 # @seealso \code{\link{print.BibEntry}}, \code{\link{BibEntry}}
 #' @importFrom tools Rd2txt_options Rd2txt Rd2HTML Rd2latex loadPkgRdMacros
-#' @S3method format BibEntry
+## @S3method format BibEntry
+#' @export
 #' @keywords internal
 format.BibEntry <- function(x, style = .BibOptions$style, .bibstyle = .BibOptions$bib.style,
                              citation.bibtex.max = getOption("citation.bibtex.max", 1), .sort = TRUE,
