+++ man/cash-.BibEntry.Rd
@@ -33,4 +33,3 @@ Other operators: \code{\link{$<-.BibEntry}},
   \code{\link{[<-.BibEntry}}, \code{\link{[[.BibEntry}},
   \code{\link{[[<-.BibEntry}}, \code{\link{c.BibEntry}}
 }
-
