+++ man/GetPubMedRelated.Rd
@@ -56,4 +56,3 @@ Other pubmed: \code{\link{GetDOIs}},
   \code{\link{ReadPubMed}}
 }
 \keyword{database}
-
