+++ man/as.BibEntry.Rd
@@ -3,6 +3,7 @@
 \name{as.BibEntry}
 \alias{as.BibEntry}
 \alias{is.BibEntry}
+\alias{is.BibEntry}
 \title{Coerce to a BibEntry object}
 \usage{
 as.BibEntry(x)
@@ -55,4 +56,3 @@ as.BibEntry(bib)
 \code{\link{BibEntry}}
 }
 \keyword{utilities}
-
