+++ NAMESPACE
@@ -51,23 +51,12 @@ export(is.BibEntry)
 export(toBiblatex)
 import(bibtex)
 import(methods)
-importFrom(RCurl,curlEscape)
-importFrom(RCurl,curlOptions)
-importFrom(RCurl,getForm)
-importFrom(RCurl,getURL)
-importFrom(RCurl,postForm)
-importFrom(XML,free)
-importFrom(XML,getNodeSet)
-importFrom(XML,xmlDoc)
-importFrom(XML,xmlGetAttr)
-importFrom(XML,xmlParse)
-importFrom(XML,xmlValue)
-importFrom(XML,xpathApply)
 importFrom(httr,GET)
 importFrom(httr,POST)
 importFrom(httr,add_headers)
 importFrom(httr,content)
 importFrom(httr,http_error)
+importFrom(httr,read_xml)
 importFrom(jsonlite,fromJSON)
 importFrom(jsonlite,toJSON)
 importFrom(lubridate,"%within%")
@@ -97,6 +86,7 @@ importFrom(tools,loadPkgRdMacros)
 importFrom(tools,parseLatex)
 importFrom(tools,toRd)
 importFrom(utils,URLdecode)
+importFrom(utils,URLencode)
 importFrom(utils,as.person)
 importFrom(utils,bibentry)
 importFrom(utils,browseURL)
@@ -105,5 +95,8 @@ importFrom(utils,head)
 importFrom(utils,person)
 importFrom(utils,tail)
 importFrom(utils,toBibtex)
+importFrom(xml2,read_xml)
+importFrom(xml2,xml_attr)
 importFrom(xml2,xml_find_all)
+importFrom(xml2,xml_get_all)
 importFrom(xml2,xml_text)
