+++ R/07makeBibLatexAuthoryear.R
@@ -552,7 +552,8 @@ fmtEprint <- switch(docstyle, html = function(paper){
         res <- paste0(paper$archiveprefix, ': ', paper$eprint, ' [', paper$eprintclass, ']')
       }else if(length(paper$primaryclass)){
         res <- paste0(paper$archiveprefix, ': ', paper$eprint, ' [', paper$primaryclass, ']')
-      }
+      }else
+        res <- paste0('eprint: ', paper$eprint)
     }else{
       res <- paste0('eprint: ', paper$eprint)
     }
@@ -588,7 +589,8 @@ fmtEprint <- switch(docstyle, html = function(paper){
         res <- paste0(paper$archiveprefix, ': ', paper$eprint, ' [', paper$eprintclass, ']')
       }else if(length(paper$primaryclass)){
         res <- paste0(paper$archiveprefix, ': ', paper$eprint, ' [', paper$primaryclass, ']')
-      }
+      }else
+        res <- paste0('eprint: ', paper$eprint)
     }else{
       res <- paste0('eprint: ', paper$eprint)
     }
@@ -612,7 +614,8 @@ fmtEprint <- switch(docstyle, html = function(paper){
         res <- paste0(paper$archiveprefix, ': ', paper$eprint, ' [', paper$eprintclass, ']')
       }else if(length(paper$primaryclass)){
         res <- paste0(paper$archiveprefix, ': ', paper$eprint, ' [', paper$primaryclass, ']')
-      }
+      }else
+        res <- paste0('eprint: ', paper$eprint)
     }else{
       res <- paste0('eprint: ', paper$eprint)
     }
