### 2017-06-01 f8daf01
correct import of read_xml
