+++ NAMESPACE
@@ -56,7 +56,6 @@ importFrom(httr,POST)
 importFrom(httr,add_headers)
 importFrom(httr,content)
 importFrom(httr,http_error)
-importFrom(httr,read_xml)
 importFrom(jsonlite,fromJSON)
 importFrom(jsonlite,toJSON)
 importFrom(lubridate,"%within%")
