### 2017-06-01 e271287
bump version, update NEWS
