+++ inst/NEWS.md
@@ -1,3 +1,10 @@
+Changes in Version 0.14.0 (2017-06-02)
+--------------------------------------------------------
+NEW FEATURES
+
+* Package now uses httr, xml2, jsonlite packages instead of RCurl, XML, RJSONIO
+for scaffolding
+
 Changes in Version 0.13.4 (2017-04-25)
 --------------------------------------------------------
 
