+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: RefManageR
-Version: 0.13.4
-Date: 2017-04-25
+Version: 0.14.0
+Date: 2017-06-02
 Title: Straightforward 'BibTeX' and 'BibLaTeX' Bibliography Management
 Authors@R: person(c("Mathew", "W."), "McLean", role = c("aut", "cre"),
     email = "mathew.w.mclean@gmail.com")
