+++ NAMESPACE
@@ -97,5 +97,4 @@ importFrom(utils,toBibtex)
 importFrom(xml2,read_xml)
 importFrom(xml2,xml_attr)
 importFrom(xml2,xml_find_all)
-importFrom(xml2,xml_get_all)
 importFrom(xml2,xml_text)
