### 2017-06-01 412ed3f
remove RCurl ref. from doc.
