### 2017-06-02 b17ae12
remove url.exists from Crossref tests
