### 2017-06-02 932ef20
check coverage for tests+vignettes+examples
