+++ .travis.yml
@@ -37,4 +37,4 @@ notifications:
     on_failure: always # default: always
 
 after_success:
- - Rscript -e 'library(covr);coveralls()'
+ - Rscript -e 'library(covr); coveralls(type = "all")'
