### 2017-06-02 59a8269
run readPDF tests
