+++ tests/test-all.R
@@ -1,3 +1,3 @@
 library(testthat)
 Sys.setenv(NOT_CRAN = "true")
-test_check("RefManageR", perl = TRUE, filter = "^(?!readPDF)[A-Za-z0-9]+$")
+test_check("RefManageR", perl = TRUE)  # , filter = "^(?!readPDF)[A-Za-z0-9]+$")
