### 2017-06-02 45fe3c4
remove duplicated formatting funs
