+++ R/CommonFormatFuns.R
@@ -1,0 +1,473 @@
+#' @keywords internal
+GetFormatFunctions <- function(docstyle = "text"){
+    collapse <- function(strings){
+      paste(strings, collapse = "\n")
+    }
+
+    labelclean <- function(prefix = NULL, suffix = NULL, style = plain){
+        f <- label(prefix, suffix, style)
+        function(s) f(cleanupLatex(s))
+    }
+
+    label <- function(prefix = NULL, suffix = NULL, style = plain){
+        force(prefix)
+        force(suffix)
+        force(style)
+        function(s) if (length(s))
+            style(paste0(prefix, collapse(s), suffix))
+    }
+
+    labelPersons <- function(prefix = NULL, suffix = NULL, style = plain){
+        force(prefix)
+        force(suffix)
+        force(style)
+        function(s) if (length(s))
+            style(paste0(prefix, authorList(s), suffix))
+    }
+
+    cleanap <- function(s){
+      if (length(s))
+        cleanupLatex(addPeriod(s))
+    }
+
+    emphclean <- function (s){
+      emph(cleanupLatex(s))
+    }
+
+    emphcleanap <- function (s){
+      emph(addPeriod(cleanupLatex(s)))
+    }
+
+    emph <- function (s){
+      if (length(s))
+        paste0("\\emph{", collapse(s), "}")
+    }
+
+    plainclean <- function (s){
+      plain(cleanupLatex(s))
+    }
+
+    plain <- function (pages){
+      if (length(pages))
+        collapse(pages)
+    }
+
+    sentence <- function (..., sep = ", "){
+        strings <- c(...)
+        if (length(strings)) {
+            addPeriod(paste(strings, collapse = sep))
+        }
+    }
+    
+    fmtEventTitle <- cleanap
+    
+    fmtIBTitle <- function(tl, stl, bib){
+      if (bib){  # bookinbook entry
+        fmtBTitle(tl, stl)
+      }else{
+        if (!is.null(stl)){
+          fmtJTitle(paste0(c(addPeriod(tl), stl), collapse =' '))
+        }else{
+          fmtJTitle(tl)
+        }
+      }
+    }
+
+    fmtAddendum <- cleanap
+    fmtAddOn <- cleanap
+    fmtHowPublished <- cleanap
+    fmtOtherField <- cleanap
+    
+    fmtJTitle <- function(title){
+      if (!is.null(title))  
+        if (grepl('[.?!]$', title, useBytes = TRUE))
+          paste0("\\dQuote{", collapse(cleanupLatex(title)), "}")
+        else paste0("\\dQuote{", collapse(cleanupLatex(title)), "}.")
+    }
+
+    fmtType <- function(type){
+      if (length(type)){
+        ind <- match(type, names(.BibEntryTypeField))
+        if (is.na(ind)){
+          cleanupLatex(type)
+        }else{
+          .BibEntryTypeField[ind]
+        }
+      }
+    }
+
+    fmtTranslator <- function(paper){
+      if (length(paper$translator))
+          paste0('Trans. ', fmtLangOrig(paper$origlanguage), ' by ',
+                 authorList(paper$translator), '.')
+    }
+
+    fmtLangOrig <- function(lang){
+      if (length(lang))
+        paste0('from the ', sub("\\b(\\w)",    "\\U\\1", lang, perl=TRUE, useBytes = TRUE))
+    }
+
+    fmtLanguage <- function(lang){
+      if (length(lang) && tolower(lang) != 'english')
+        addPeriod(sub("\\b(\\w)",    "\\U\\1", lang, perl=TRUE, useBytes = TRUE))
+    }
+
+    fmtSeries <- label(prefix = '. ')
+
+    fmtPubstate <- function(ps){
[…]
