+++ R/07makeBibLatexAuthoryear.R
@@ -1,1119 +1,721 @@
 #' @keywords internal
 MakeAuthorYear <- function(docstyle = "text"){
-##################################################################
-## Formatting functions
+  with(GetFormatFunctions(docstyle), {
+    ##################################################################
+    ## Formatting functions
 
-collapse <- function(strings){ 
-  paste(strings, collapse = "\n")
-}
-
-collapseF <- function(strings){ 
-  out <- paste(strings, collapse = "\n")
-  gsub('\\.((\\n)?,)', '\\1', out, useBytes = TRUE)
-}
-
-labelclean <- function(prefix = NULL, suffix = NULL, style = plain){
-    f <- label(prefix, suffix, style)
-    function(s) f(cleanupLatex(s))
-}
-
-label <- function(prefix = NULL, suffix = NULL, style = plain){
-    force(prefix)
-    force(suffix)
-    force(style)
-    function(s) if (length(s)) 
-        style(paste0(prefix, collapse(s), suffix))
-}
-
-labelPersons <- function(prefix = NULL, suffix = NULL, style = plain){
-    force(prefix)
-    force(suffix)
-    force(style)
-    function(s) if (length(s)) 
-        style(paste0(prefix, authorList(s), suffix))
-}
-
-
-emphclean <- function (s){
-  emph(cleanupLatex(s))
-}
-
-emphcleanap <- function (s){
-  emph(addPeriod(cleanupLatex(s)))
-}
-
-emph <- function (s){
-  if (length(s)) 
-    paste0("\\emph{", collapse(s), "}")
-}
-
-plainclean <- function (s){
-  plain(cleanupLatex(s))
-}
-
-cleanap <- function(s){
-  cleanupLatex(addPeriod(s))
-}
-
-plain <- function (pages){
-  if (length(pages)) 
-    collapse(pages)
-}
-
-sentence <- function (..., sep = ", "){
-    strings <- c(...)
-    if (length(strings)) {
-        addPeriod(paste(strings, collapse = sep))
+    collapseF <- function(strings){ 
+      out <- paste(strings, collapse = "\n")
+      gsub('\\.((\\n)?,)', '\\1', out, useBytes = TRUE)
     }
-}
 
-addPeriod <- function (string){ 
-  sub("([^.?!])$", "\\1.", string, useBytes = TRUE)
-}
-
-addPeriodTitle <- function (string){ 
-  sub("([^.?!])$", "\\dQuote{\\1.}", string, useBytes = TRUE)
-}
-
-authorList <- function(aut){
-    names <- sapply(aut, shortName)
-    if (length(names) > 1L){ 
-        result <- paste(paste(names[-length(names)], collapse = ", "), 
-            "and", names[length(names)])
-    }else{
-      result <- names
+    addPeriodTitle <- function (string){ 
+      sub("([^.?!])$", "\\dQuote{\\1.}", string, useBytes = TRUE)
     }
-    result
-}
 
-fmtSingleEditor <- function(nom, job, prefix = NULL, suffix = '.'){
-  if (length(nom)){
-    if(length(job)){
-      res <- paste0(switch(tolower(job), 'compiler' = 'Comp. by ', 'editor' = 'Ed. by ', 
-                    'founder' = 'Found. by ', 'continuator' = 'Cont. by ', 'redactor' = 'Red. by ', 
-                    'reviser' = 'Rev. by ', 'collaborator' = 'In collab. with ', job), nom)
-    }else{
-      res <- paste0('Ed. by ', nom)
-    }
-    paste0(prefix, res, suffix)
-  }
-}
-
-shortNameLF <- function(pers){
-  fam <- pers$family
-  lfam <- length(fam)
-  if (lfam) {
-    von <- lfam > 1L && substr(fam[1L], 1L, 1L) %in% letters
-    if (von){
[…]

[…]
