### 2017-06-02 10935a2
add back in fmtBAuthor
