### 2017-06-02 a4800ed
pass DateFormatter to GetFormatFunctions to avoid check NOTE no vis. binding
