+++ R/05makeBibLatex.R
@@ -1,9 +1,27 @@
 #' @keywords internal
 MakeBibLaTeX <- function(docstyle = "text", authortitle = FALSE) {
-    penv <- GetFormatFunctions(docstyle)
-    assign("authortitle", authortitle, penv)
+  DF <- function(dat, other = FALSE){
+    if (!is.null(dat)){
+      if (other){
+          fmt <- switch(as.character(attr(dat, 'day.mon')), '2' = '%m/%d/%Y',
+                        '1' ='%m/%Y', '0' ='%Y')
+      }else{
+          fmt <- switch(as.character(attr(dat, 'day.mon')), '2' = '%b. %d, %Y',
+                        '1' = '%b. %Y', '0' = '%Y')
+      }
+      if (is.interval(dat)){
+        begind <- int_start(dat)
+        endd <- int_end(dat)
+        paste(format(begind, fmt), format(endd, fmt), sep ='-')
+      }else{
+        format(dat, fmt)
+      }
+    }
+  }
+
+  penv <- GetFormatFunctions(docstyle, DF)
+  assign("authortitle", authortitle, penv)
   with(penv, {
-    force(authortitle)
     ##################################################################
     ## Formatting functions
 
@@ -13,719 +31,741 @@ MakeBibLaTeX <- function(docstyle = "text", authortitle = FALSE) {
                                           draft = paste0('\\bold{', attr(paper, 'key'), '}'),
                                           NULL)
 
-fmtPrefix <- switch(docstyle, html = function(paper){
-  res <- fmtPrefixSimple(paper)
-  if (length(res)){
-    key <- attr(paper, "key")
-    ind <- .cites$indices[key]
-    key <- gsub("[^_a-zA-Z0-9-]", "", key, useBytes = TRUE)
-    res <- if (!is.na(ind) && ind)
-      paste0("\\code{", key, "}\\href{#cite-", key, "}{", res, "}")
-    else res
+  fmtPrefix <- switch(docstyle, html = function(paper){
+    res <- fmtPrefixSimple(paper)
+    if (length(res)){
+      key <- attr(paper, "key")
+      ind <- .cites$indices[key]
+      key <- gsub("[^_a-zA-Z0-9-]", "", key, useBytes = TRUE)
+      res <- if (!is.na(ind) && ind)
+        paste0("\\code{", key, "}\\href{#cite-", key, "}{", res, "}")
+      else res
+    }
+    res
+  }, markdown = function(paper){
+    res <- fmtPrefixSimple(paper)
+    if (length(res)){
+      key <- attr(paper, "key")
+      ind <- .cites$indices[key]
+      key <- gsub("[^_a-zA-Z0-9-]", "", key, useBytes = TRUE)
+      res <- if (!is.na(ind) && ind)
+        paste0("<a name=bib-", key, "></a>[", res, "](#cite-", key, ")")
+      else res
+    }
+    res
+  }, fmtPrefixSimple)
+
+
+
+  fmtNumPre <- function(doc){
+    res <- doc$shorthand
+    if (!length(res))
+      res <- doc$.index
+    res
   }
-  res
-}, markdown = function(paper){
-  res <- fmtPrefixSimple(paper)
-  if (length(res)){
-    key <- attr(paper, "key")
-    ind <- .cites$indices[key]
-    key <- gsub("[^_a-zA-Z0-9-]", "", key, useBytes = TRUE)
-    res <- if (!is.na(ind) && ind)
-      paste0("<a name=bib-", key, "></a>[", res, "](#cite-", key, ")")
-    else res
+
+  sentenceP <- function (..., pgs = NULL, tp = NULL, sep = ", "){
+    strings <- c(...)
+    res <- NULL
[…]
