### 2017-06-02 3af074a
remove dup. key sorting funs
