### 2017-06-02 9ccfb52
put all key sorting funs in common format fun env
