### 2017-06-02 a082fd9
construct env only once to get sort key funs
