### 2017-06-02 b9f9cfa
move another sort key fun
