+++ R/CommonFormatFuns.R
@@ -494,6 +494,23 @@ GetFormatFunctions <- function(docstyle = "text", DateFormatter){
       }
       result
     }
+    
+    sortKeysV <- function(bib){
+      result <- numeric(length(bib))
+      for (i in seq_along(bib)){
+        res <- bib[[i]]$volume
+        if (!length(res)){
+          res <- "0000"
+        }else{
+          tmp <- suppressWarnings(as.numeric(res))
+          res <- if (!is.na(tmp))
+                     sprintf("%04d", tmp)
+                 else res
+        }
+        result[i] <- res
+      }
+      result
+    }
 
     sortKeysLA <- function(bib, yrs){
         result <- character(length(bib))
@@ -545,6 +562,23 @@ GetFormatFunctions <- function(docstyle = "text", DateFormatter){
         res
       }
     }
-    
+
+    GetLastNames <- function(bib){
+        result <- character(length(bib))
+        for (i in seq_along(bib)) {
+          authors <- paste0(unlist(bib[[i]]$author$family), collapse = '')
+          if (authors == "")
+            authors <- paste0(unlist(bib[[i]]$editor$family), collapse = '')
+          if (authors == "")
+            authors <- paste0(unlist(bib[[i]]$translator$family), collapse = '')
+          if (authors == "")
+            authors <- cleanupLatex(bib[[i]]$title)
+          if (!length(authors))
+            authors <- ""
+          result[i] <- authors
+        }
+        result
+    }
+        
     environment()
 }
