### 2017-06-02 1705d03
fix yaml style formatting
