### 2017-06-02 b5ed5ae
mention yaml style in docs
