### 2017-06-02 0008410
add tests more tests of printing
