### 2017-06-02 075599a
update NEWS and version
