+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: RefManageR
-Version: 0.14.1
+Version: 0.14.3
 Date: 2017-06-02
 Title: Straightforward 'BibTeX' and 'BibLaTeX' Bibliography Management
 Authors@R: person(c("Mathew", "W."), "McLean", role = c("aut", "cre"),
