### 2017-06-02 5498e23
add noRd tag to roxygen
