### 2017-06-09 6c65675
Bibtex months are expected in English
Most bibtex entries are in English, so we should expect the months to be in English as well. `lubridate::parse_date_time` uses the current locale instead of the English "C" locale.
