### 2017-06-12 84c350b
fix typos in Rd files
