+++ R/BibEntryListExtract.R
@@ -3,7 +3,7 @@
 #' Operator for extracting BibEntry objects by index.
 #' 
 #' @param x a BibEntry object
-#' @param i numeric indices of entries to exctract, or a character vector of keys corresponding to the entries to be
+#' @param i numeric indices of entries to extract, or a character vector of keys corresponding to the entries to be
 #' extracted.
 #' @param drop logical, should attributes besides class be dropped from result?
 #' @method [[ BibEntry
