+++ man/BibOptions.Rd
@@ -56,7 +56,7 @@ recurring author and editor names are replaced with \dQuote{---} when printing.
 \code{\link{sort.BibEntry}}
 \item \code{max.names} - numeric; maximum number of names to display before using \dQuote{et al.} when formatting and printing name
 list fields.  This is also the minimum number of names that will be displayed if \dQuote{et al.} is used 
-(minnames package option in Biblatex) 
+(\sQuote{minnames} package option in Biblatex) 
 \item \code{no.print.fields} character vector; fields that should not be printed, 
 e.g., doi, url, isbn, etc.
 \item \code{style} - character string naming the printing style.  Possible values are 
