### 2017-06-14 #269 @karthik
Have sent a request to @cboettig for a review. Will update thread soon.
