### 2017-06-14 #269 @karthik
@cboettig Please ignore this request. I see that @noamross has already pinged you at ropensci onboarding. Will place this issue into holding till the review is complete there. @noamross Please ping when that is done.
