### 2017-06-15 c2eabec
SearchDOIText always returns chars
