### 2017-06-15 e3370f2
remove final mentions of RCurl in tests
