+++ tests/testthat/test-methods.R
@@ -52,7 +52,7 @@ test_that("head and tail", {
 test_that("open", {
     skip_on_cran()
     skip_on_travis()
-    open(as.BibEntry(citation("RCurl")))  # URL
+    open(as.BibEntry(citation("httr")))  # URL
 
     testbib <- ReadBib(system.file("REFERENCES.bib", package="bibtex"))
     testbib$file <- file.path(R.home("doc/manual"), "R-intro.pdf")
