### 2017-06-15 33c414e
error instead of warn if try to combine non-BibEntry obj
