### 2017-06-15 2580463
test combine operator with BibEntry object and length zero objects
