### 2017-06-15 79585c8
switch sapply to vapply
