+++ R/BibEntryListExtract.R
@@ -34,7 +34,7 @@
   if (is.character(i) && is.null(names(x))) 
     names(x) <- .BibEntry_get_key(x)
   y <- x[i]
-  if (!all(ok <- sapply(y, length) > 0L)) {
+  if (!all(ok <- vapply(y, length, 0L) > 0L)) {
     warning("subscript out of bounds")
     y <- y[ok]
   }

[…]
