### 2017-06-15 f6d95a3
remove unknown char from bib file
