+++ inst/Bib/RJC.bib
@@ -65,7 +65,7 @@
 }
 
 @Article{tooze2013measurement,
-  title = {A Measurement Error Model for Physical Activity Level as Measured by a Questionnaire With Application to the 1999–2006 NHANES Questionnaire},
+  title = {A Measurement Error Model for Physical Activity Level as Measured by a Questionnaire With Application to the 1999--2006 NHANES Questionnaire},
   author = {J A Tooze and R P Troiano and R J Carroll and A J Moshfegh and L S Freedman},
   journal = {American journal of epidemiology},
   number = {11},
