### 2017-06-15 c454d06
update NEWS, version
