### 2017-06-15 0a99cb0
fix sapply to vapply chg in sort tests
