+++ R/toBibtex.R
@@ -3,19 +3,23 @@
 #' @importFrom utils toBibtex
 #' @export
 toBibtex.BibEntry <- function(object,
-                              note.replace.field = c('urldate', "pubsate", "addendum"),
+                              note.replace.field = c('urldate', "pubsate",
+                                                     "addendum"),
                               extra.fields = NULL, ...){
   format_bibentry1 <- function(object){
     object <- unclass(object)[[1L]]
     bibtype <- tolower(attr(object, "bibtype"))
     obj.names <- names(object)
     if ("author" %in% obj.names)
-      object$author <- encoded_text_to_latex(format_author(object$author), "UTF-8")
+      object$author <- encoded_text_to_latex(format_author(object$author),
+                                             "UTF-8")
     if ("editor" %in% obj.names)
-      object$editor <- encoded_text_to_latex(format_author(object$editor), "UTF-8")
+      object$editor <- encoded_text_to_latex(format_author(object$editor),
+                                             "UTF-8")
     # see 2.3 Usage Notes p. 28
     
-    if (bibtype == "article" && 'journaltitle' %in% obj.names  && is.null(object$journal))
+    if (bibtype == "article" && 'journaltitle' %in% obj.names  &&
+        is.null(object$journal))
       object$journal <- object$journaltitle
     
     if ("location" %in% obj.names  && is.null(object$address))
@@ -37,7 +41,8 @@ toBibtex.BibEntry <- function(object,
       }
     }
     
-    if ("institution" %in% obj.names && bibtype == 'thesis' && is.null(object$school)){
+    if ("institution" %in% obj.names && bibtype == 'thesis' &&
+        is.null(object$school)){
       object$school <- object$institution
       object$institution <- NULL 
     }
@@ -61,7 +66,8 @@ toBibtex.BibEntry <- function(object,
           if (note.replace.field[i] == 'urldate'){
             fDate <- try(ProcessDate(object$urldate, NULL), TRUE)
             if (!is.null(fDate) && !inherits(fDate, 'try-error')){
-              object$note <- paste0('Last visited on ', MakeBibLaTeX()$DateFormatter(fDate, TRUE))  
+              object$note <- paste0('Last visited on ',
+                                    MakeBibLaTeX()$DateFormatter(fDate, TRUE))
             }else{
               object$note <- paste0('Last visited on ', object$urldate)
             }
@@ -85,18 +91,24 @@ toBibtex.BibEntry <- function(object,
     
     pos <- match(bibtype, tolower(names(BibTeX_entry_field_db)))
     if (is.na(pos)){
-      bibtype <- switch(bibtype, "mvbook" = "Book", "bookinbook" = "InBook", "suppbook" = "InBook",
-                        "collection" = "Book", "mvcollection" = "Book", "suppcollection" = "InCollection",
-                        "reference" = "Book", "mvreference" = "Book", "inreference" = "InBook",
-                        "report" = "TechReport", "proceedings" = "Book", "mvproceedings" = "Book",
-                        "periodical" = "Book", "suppperiodical" = "InBook", "patent" = "TechReport", "Misc")
+        bibtype <- switch(bibtype, "mvbook" = "Book", "bookinbook" = "InBook",
+                          "suppbook" = "InBook", "collection" = "Book",
+                          "mvcollection" = "Book",
+                          "suppcollection" = "InCollection",
+                          "reference" = "Book", "mvreference" = "Book",
+                          "inreference" = "InBook", "report" = "TechReport",
+                          "proceedings" = "Book", "mvproceedings" = "Book",
+                          "periodical" = "Book", "suppperiodical" = "InBook",
+                          "patent" = "TechReport", "Misc")
     }else{
       bibtype <- names(BibTeX_entry_field_db)[pos]
[…]
