### 2017-06-19 f8a1957
add appveyor yaml
