+++ .appveyor.yml
@@ -1,0 +1,56 @@
+init:
+  ps: |
+        $ErrorActionPreference = "Stop"
+        Invoke-WebRequest http://raw.github.com/krlmlr/r-appveyor/master/scripts/appveyor-tool.ps1 -OutFile "..\appveyor-tool.ps1"
+        Import-Module '..\appveyor-tool.ps1'
+install:
+  ps: Bootstrap
+
+
+cache:
+  - C:\RLibrary
+
+environment:
+  matrix:
+    - R_VERSION: devel
+    - R_VERSION: release
+      RTOOLS_VERSION: 32
+      CRAN: http://cran.rstudio.com
+    - POPPLERQT5FILE: 'dev\windows\appveyor\poppler-qt5-debug.zip'
+  global:
+    NOT_CRAN: true
+
+build_script:
+  - travis-tool.sh install_deps
+  - travis-tool.sh install_github jimhester/covr
+#  - travis-tool.sh install_github rstats-db/RSQLite
+
+test_script:
+  - travis-tool.sh run_tests
+
+on_failure:
+  - travis-tool.sh dump_logs
+  - 7z a failure.zip *.Rcheck\*
+  - appveyor PushArtifact failure.zip
+
+on_success:
+  - R -e 'covr::coveralls(type = "all")'
+
+artifacts:
+  - path: '*.Rcheck\**\*.log'
+    name: Logs
+
+  - path: '*.Rcheck\**\*.out'
+    name: Logs
+
+  - path: '*.Rcheck\**\*.fail'
+    name: Logs
+
+  - path: '*.Rcheck\**\*.Rout'
+    name: Logs
+
+  - path: '\*_*.tar.gz'
+    name: Bits
+
+  - path: '\*_*.zip'
+    name: Bits
