### 2017-06-19 b0bb156
try again for poppler
