+++ .appveyor.yml
@@ -13,10 +13,11 @@ cache:
 environment:
   matrix:
     - R_VERSION: devel
+      POPPLERQT5FILE: 'dev\windows\appveyor\poppler-qt5-debug.zip'
     - R_VERSION: release
       RTOOLS_VERSION: 32
       CRAN: http://cran.rstudio.com
-    - POPPLERQT5FILE: 'dev\windows\appveyor\poppler-qt5-debug.zip'
+      POPPLERQT5FILE: 'dev\windows\appveyor\poppler-qt5-debug.zip'
   global:
     NOT_CRAN: true
 
