### 2017-06-19 eff62db
remove covr
