+++ .appveyor.yml
@@ -1,7 +1,7 @@
 init:
   ps: |
         $ErrorActionPreference = "Stop"
-        Invoke-WebRequest http://raw.github.com/mwmclean/r-appveyor/master/scripts/appveyor-tool.ps1 -OutFile "..\appveyor-tool.ps1"
+        Invoke-WebRequest http://raw.github.com/krlmlr/r-appveyor/master/scripts/appveyor-tool.ps1 -OutFile "..\appveyor-tool.ps1"
         Import-Module '..\appveyor-tool.ps1'
 install:
   ps: Bootstrap
@@ -37,7 +37,7 @@ on_failure:
   - appveyor PushArtifact failure.zip
 
 # on_success:
-#   - R -e "covr::coveralls(type = 'all')"
+#  - R -e "covr::coveralls(type = 'all')"
 
 artifacts:
   - path: '*.Rcheck\**\*.log'
