### 2017-06-24 d08d850
replace '\&' with '&' when cleaning LaTeX for printing; closes #30
