### 2017-06-24 4ceda23
skip all pubmed tests if can't connect
