### 2017-06-28 ec00b5a
add contributing guidelines
