+++ CONTRIBUTING.md
@@ -1,0 +1,24 @@
+# Contributing to RefManageR
+
+The following is modified from the [plotly contributing
+guidelines](https://github.com/ropensci/plotly/blob/master/CONTRIBUTING.md).
+
+## Opening issues
+
+When reporting an issue please provide a reproducible example.  You
+may find the package [reprex](https://github.com/tidyverse/reprex)
+useful for this.
+
+## Development Guidelines
+
+If you'd like to contribute changes to RefManageR, see [the GitHub
+flow](https://guides.github.com/introduction/flow/index.html) for
+proposing, submitting, reviewing, and accepting changes. If you
+haven't done this before, Hadley Wickham provides a nice overview of
+git (<http://r-pkgs.had.co.nz/git.html>), as well as best practices
+for submitting pull requests
+(<http://r-pkgs.had.co.nz/git.html#pr-make>). I also recommend using
+his style guide when writing code
+(<http://adv-r.had.co.nz/Style.html>).  If your pull request fixes a
+bug, or implements a new feature, it's a good idea to write a test
+(<http://r-pkgs.had.co.nz/tests.html>) to demonstrate it's working.
