### 2017-06-30 e3f8ef1
royxgen use @export instead of @S3method
