### 2017-06-30 2fa3841
remove NOTE thanks to bibtex update
