+++ NAMESPACE
@@ -20,6 +20,7 @@ S3method(print,BibEntry)
 S3method(sort,BibEntry)
 S3method(tail,BibEntry)
 S3method(toBibtex,BibEntry)
+S3method(toRd,BibEntry)
 S3method(unlist,BibEntry)
 export(AutoCite)
 export(BibEntry)
@@ -49,7 +50,7 @@ export(as.BibEntry)
 export(fields)
 export(is.BibEntry)
 export(toBiblatex)
-import(bibtex)
+importFrom(bibtex,do_read_bib)
 importFrom(httr,GET)
 importFrom(httr,POST)
 importFrom(httr,add_headers)
