+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: RefManageR
-Version: 0.14.11
+Version: 0.14.12
 Title: Straightforward 'BibTeX' and 'BibLaTeX' Bibliography Management
 Authors@R: person(c("Mathew", "W."), "McLean", role = c("aut", "cre"),
     email = "mathew.w.mclean@gmail.com")
@@ -25,7 +25,7 @@ Imports:
     plyr,
     tools,
     httr,
-    bibtex,
+    bibtex (>= 0.4.1),
     lubridate (>= 1.5.0),
     stringr,
     methods
