### 2017-06-30 bd867fa
install github version of bibtex
