+++ .appveyor.yml
@@ -24,9 +24,11 @@ environment:
 
 build_script:
   - travis-tool.sh install_deps
+#  - travis-tool.sh install_github romainfrancois/bibtex
 #  - Rscript -e "install.packages('covr', repos = 'http://cran.rstudio.com')"
 #  - travis-tool.sh install_github jimhester/covr
 #  - travis-tool.sh install_github rstats-db/RSQLite
+
 
 test_script:
   - travis-tool.sh run_tests
