### 2017-06-30 372df23
fix bibtex github install
