### 2017-07-03 8678ee6
don't bother with github version of bibtex
