+++ .appveyor.yml
@@ -24,8 +24,8 @@ environment:
 
 build_script:
   - travis-tool.sh install_deps
-  - travis-tool.sh install_r devtools
-  - Rscript -e 'devtools::install_github("romainfrancois/bibtex")'
+#  - travis-tool.sh install_r devtools
+#  - Rscript -e 'devtools::install_github("romainfrancois/bibtex")'
 #  - travis-tool.sh install_github romainfrancois/bibtex
 #  - Rscript -e "install.packages('covr', repos = 'http://cran.rstudio.com')"
 #  - travis-tool.sh install_github jimhester/covr
