### 2017-07-04 13c6776
CMD check doesn't like coveralls url
