### 2017-07-18 4384f73
check if month is empty text (e.g. character(0)) when processing dates
