### 2017-07-18 4991545
update Rd example
