### 2017-07-18 856e57d
try different header when GETting bibtex
