### 2017-07-18 56b17e5
add httr:: to Rd examples
