+++ man/ReadPubMed.Rd
@@ -52,7 +52,7 @@ and the abstract will be returned in the field \dQuote{abstract}, if they
 are available.
 }
 \examples{
-if (interactive() && !http_error("https://eutils.ncbi.nlm.nih.gov/"))
+if (interactive() && !httr::http_error("https://eutils.ncbi.nlm.nih.gov/"))
   ReadPubMed(query = "raymond carroll measurement error", retmax = 5, mindate = 1990)
 }
 \references{
