### 2017-07-18 30d0063
add Rd examples for NoCite and PrintBibliography
