### 2017-07-18 bc384d4
add Rd examples
