### 2017-07-18 b2353f1
decode URL returned by DOI System API
