### 2017-07-18 32d7cc7
remove use of deprec. testthat fun
