### 2017-07-18 52a5366
fix links from bibliography to citations in vignettes
