### 2017-07-24 9b391c0
use http_error not try for error handling with API funs
