+++ NAMESPACE
@@ -56,6 +56,7 @@ importFrom(httr,POST)
 importFrom(httr,add_headers)
 importFrom(httr,content)
 importFrom(httr,http_error)
+importFrom(httr,status_code)
 importFrom(jsonlite,fromJSON)
 importFrom(jsonlite,toJSON)
 importFrom(lubridate,"%within%")
