### 2017-07-24 d54e916
refactor to reduce cyclomatic complexity
