### 2017-07-24 8bf9a79
refactor to reduce cyclomatic complexity
