### 2017-07-24 3fc0bc1
replace another try with http_error
