+++ R/ReadCrossRef.R
@@ -217,24 +217,21 @@ GetCrossRefBibTeX <- function(doi, tmp.file){
     ## temp <- try(getURLContent(url=doi,
     ##  .opts = curlOptions(httpheader = c(Accept = "application/x-bibtex"),
     ##  followLocation=TRUE)), TRUE)
-    temp <- try(GET(doi, config = list(followlocation = TRUE),
-                    add_headers(Accept = "application/x-bibtex")), TRUE)
-    temp <- try(content(temp, as = "text", encoding = "UTF-8"), TRUE)
+    temp <- GET(doi, config = list(followlocation = TRUE),
+                    add_headers(Accept = "application/x-bibtex"))
+    parsed <- content(temp, as = "text", encoding = "UTF-8")
     ## if(is.raw(temp))
     ##     temp <- rawToChar(temp)
-    if (inherits(temp, "try-error") ||
-        temp[1] == "<h1>Internal Server Error</h1>" ||
-        !grepl("^[[:space:]]*@", temp, useBytes = TRUE)){
+    if (http_error(temp) ||
+        !grepl("^[[:space:]]*@", parsed, useBytes = TRUE)){
         ## last one for occasional non-bibtex returned by CrossRef
 
         ## try different header if first one fails
-        temp <- try(GET(doi, config = list(followlocation = TRUE),
-                    add_headers(Accept = "text/bibliography; style=bibtex")),
-                    TRUE)
-        temp <- try(content(temp, as = "text", encoding = "UTF-8"), TRUE)        
-        if (inherits(temp, "try-error") ||
-            temp[1] == "<h1>Internal Server Error</h1>" ||
-            !grepl("^[[:space:]]*@", temp, useBytes = TRUE)){
+        temp <- GET(doi, config = list(followlocation = TRUE),
+                    add_headers(Accept = "text/bibliography; style=bibtex"))
+        parsed <- content(temp, as = "text", encoding = "UTF-8")
+        if (http_error(temp) ||
+            !grepl("^[[:space:]]*@", parsed, useBytes = TRUE)){
             doi.text <- sub("^https?://(dx[.])?doi.org/", "", doi)
             message(gettextf("Server error for doi %s, you may want to %s",
                              dQuote(doi.text),
@@ -243,10 +240,10 @@ GetCrossRefBibTeX <- function(doi, tmp.file){
         }
     }
 
-    temp <- gsub("&amp;", "&", temp, useBytes = TRUE)
+    parsed <- gsub("&amp;", "&", parsed, useBytes = TRUE)
     ## Crossref uses data type for some entries
-    temp <- sub("^@[Dd]ata", "@online", temp, useBytes = TRUE)
-    write(temp, file = tmp.file, append=TRUE)
+    parsed <- sub("^@[Dd]ata", "@online", parsed, useBytes = TRUE)
+    write(parsed, file = tmp.file, append = TRUE)
     return(0L)
 }
 
