### 2017-07-24 3513574
more refactoring
