### 2017-07-24 eb1a768
fix ResolveCrossref typo
