### 2017-07-25 8413012
try CRAN version of covr
