### 2017-07-25 d5c5424
refactor to reduce cyclomatic complexity
