+++ R/rmdCite.R
@@ -86,38 +86,15 @@ Cite <- function(bib, ..., textual = FALSE, before = NULL, after = NULL,
     if (identical(class(bib), "bibentry"))
       bib <- as.BibEntry(bib)
 
-    result <- with(BibOptions(), {
+    with(BibOptions(), {
       style <- .BibEntry_match_format_style(style)
       papers <- suppressMessages(do.call(`[.BibEntry`, list(x = bib, ...)))
       keys <- unlist(papers$key)
       if (!length(papers))
         return("")
       if (cite.style == "pandoc"){
-        result <- paste0(paste0("@", names(papers)),
-                         collapse = paste0(bibpunct[5L], " "))
-        result <- paste0(before, result, after)
-        if (textual)
-          result <- paste0(bibpunct[3L], result,
-                             bibpunct[4L])
-        AddCite(keys, FALSE)
-        result
+        MakePandocCitation(papers, keys, textual, bibpunct, before, after)  
       }else{
-        shortName <- function(person){
-            if (length(person$family))
-                paste(cleanupLatex(person$family), collapse = " ")
-            else paste(cleanupLatex(person$given), collapse = " ")
-        }
-        authorList <- function(paper){
-          names <- vapply(paper$author, shortName, "")
-          if (!length(names))
-            names <- vapply(paper$editor, shortName, "")
-          if (!length(names))
-            names <- paper$label
-          if (!length(names))
-            names <- vapply(paper$translator, shortName, "")
-          names
-        }
-
         numeric <- "numeric" %in% cite.style
         alphabetic <- "alphabetic" %in% cite.style
         if (cite.style != .cites$sty)
@@ -219,53 +196,111 @@ Cite <- function(bib, ..., textual = FALSE, before = NULL, after = NULL,
         }else {
             result <- paste0(auth, bibpunct[6L], " ", year)
         }
-        if (make.hyper){
-            url <- switch(hyperlink, to.bib = paste0("#bib-",
-                                                     gsub("[^_a-zA-Z0-9-]", "",
-                                                          keys,
-                                                          useBytes = TRUE)),
-                       to.doc = vapply(papers, GetURL, "", 
-                                       flds = c("url", "eprint", "doi"),
-                                       to.bib = TRUE),
-                       hyperlink)
-          if (style == "html"){
-            new.links <- if (any(first))
-                             paste(paste("<a id='cite-", gsub("[^_a-zA-Z0-9-]",
-                                                              "", keys[first],
-                                                              useBytes = TRUE),
-                                         "'></a>", sep = ""), collapse = "")
-                         else ""
-            result <- if (numeric && super && textual)
-                        paste0(auth, "<sup><a href='", url, "'>", result,
-                               "</a></sup>")
-                      else paste0("<a href='", url, "'>", result, "</a>")
-          }else if(style == "markdown"){
-            new.links <- if(any(first))
-                             paste(paste("<a name=cite-",
-                                         gsub("[^_a-zA-Z0-9-]", "",
-                                              keys[first], useBytes = TRUE),
-                                         "></a>", sep = ""), collapse = "")
-            else ""
-            result <- if (numeric && super && textual)
-                        paste0(auth, "^[", result, "](", url, ")")
-                      else paste0("[", result, "](", url, ")")
-          }
-        }
-        result <- paste(result, collapse = paste0(bibpunct[5L], " "))
-        if (!textual && (numeric || alphabetic)) {
-          result <- paste0(bibpunct[3L], before, result, after,
-                           bibpunct[4L])
-          if (super && numeric)
-            result <- paste0("^{", result, "}")
[…]
