### 2017-07-25 5bb35a9
rm superfluous msg from add operator
