+++ R/BibEntryAddOp.R
@@ -81,7 +81,8 @@
       possible.dup <- unlist(e2$bibtype) %in% unlist(e1$bibtype)
     }
     remain.dup.f <- setdiff(fields.to.check, c('bibtype', 'key'))
-    if (length(remain.dup.f)){
+    
+    if (length(remain.dup.f) && length(dup.ind <- possible.dup)){
       dup.ind <- vapply(unclass(e2[possible.dup]),
                       function(x, y, flds){
                         x <- x[flds]
@@ -93,8 +94,6 @@
                     }, FALSE, y = unclass(e1), flds = remain.dup.f)
       if (length(dup.ind))
           dup.ind <- which(dup.ind)
-    }else{
-      dup.ind <- possible.dup
     }
     if (length(dup.ind) == length(e2)){
       message('Only duplicates in second BibEntry object')
