### 2017-07-25 4dcbfd8
improve API query error handling
