+++ R/GetDOIs.R
@@ -37,10 +37,11 @@ GetDOIs <- function(bib){
     json.res <- httr::POST("http://search.crossref.org/links", body = json.bib,
                            config = list(add_headers = headers),
                            encode = "json")
-    if (http_error(json.res) ||
+    status <- status_code(json.res)
+    if (status != 200 ||
         !(json.res <- content(json.res, type = "application/json",
                               encoding = "UTF-8"))[[2L]])
-      message("Failed to retrieve any DOIs.")
+        message(gettextf("Failed to retrieve any DOIs [%s].", status))
     else{
       matches <- vapply(json.res[[1L]], "[[", FALSE, "match")
       if (!any(matches))
