+++ NAMESPACE
@@ -58,6 +58,7 @@ importFrom(httr,content)
 importFrom(httr,http_error)
 importFrom(httr,modify_url)
 importFrom(httr,status_code)
+importFrom(httr,stop_for_status)
 importFrom(jsonlite,fromJSON)
 importFrom(jsonlite,toJSON)
 importFrom(lubridate,"%within%")
