### 2017-07-27 d699094
better error msg if no poppler
