+++ inst/NEWS.md
@@ -1,4 +1,4 @@
-Changes in Version 0.14.19 (2017-07-26)
+Changes in Version 0.14.20 (2017-07-27)
 --------------------------------------------------------
 
 NEW FEATURES
