### 2017-07-27 4f9a118
Add rOpenSci stuff
