+++ README.md
@@ -1,10 +1,11 @@
 RefManageR
 ========
-[![](https://travis-ci.org/mwmclean/RefManageR.svg?branch=master)](https://travis-ci.org/mwmclean/RefManageR/)
-[![AppVeyor Build Status](http://ci.appveyor.com/api/projects/status/github/mwmclean/RefManageR?branch=master&svg=true)](http://ci.appveyor.com/project/mwmclean/RefManageR)
-[![Coverage Status](https://coveralls.io/repos/mwmclean/RefManageR/badge.svg?branch=master)](https://coveralls.io/r/mwmclean/RefManageR?branch=master)
+[![](https://travis-ci.org/ropensci/RefManageR.svg?branch=master)](https://travis-ci.org/ropensci/RefManageR/)
+[![AppVeyor Build Status](http://ci.appveyor.com/api/projects/status/github/ropensci/RefManageR?branch=master&svg=true)](http://ci.appveyor.com/project/ropensci/RefManageR)
+[![Coverage Status](https://coveralls.io/repos/ropensci/RefManageR/badge.svg?branch=master)](https://coveralls.io/r/ropensci/RefManageR?branch=master)
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/RefManageR)](https://cran.r-project.org/package=RefManageR)
 [![CRAN_Download_Badge](http://cranlogs.r-pkg.org/badges/RefManageR)](https://cran.r-project.org/package=RefManageR)
+[![](https://ropensci.org/badges/120_status.svg)](https://github.com/ropensci/onboarding/issues/120)
 
 `RefManageR` provides tools for importing and working with
 bibliographic references.  It greatly enhances the `bibentry` class by
@@ -20,7 +21,7 @@ functions for citing and generating a bibliography with hyperlinks for
 documents prepared with `RMarkdown` or `RHTML`.
 
 Please see the [vignette](https://arxiv.org/pdf/1403.2036v1)
-for an introduction and [NEWS](https://github.com/mwmclean/RefManageR/blob/master/inst/NEWS.md)
+for an introduction and [NEWS](https://github.com/ropensci/RefManageR/blob/master/inst/NEWS.md)
 for the latest changes.
 
 Using the latest version from `Github` requires first installing the newest version of the `bibtex` package
@@ -29,5 +30,6 @@ from `GitHub`:
 ```
 install.packages("devtools")
 devtools::install_github("romainfrancois/bibtex")
-devtools::install_github("mwmclean/RefManageR")
+devtools::install_github("ropensci/RefManageR")
 ```
+[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
