### 2017-07-27 3101e09
remove mention of bibtex package
