### 2017-07-27 a27c4b8
mention rOpenSci
