### 2017-07-27 #269 @noamross
@arfon This package has been approved through ropensci onboarding and the review thread is here: https://github.com/ropensci/onboarding/issues/119
