### 2017-07-27 #338 @arfon
rOpenSci onboarding thread: https://github.com/ropensci/onboarding/issues/119
