### 2017-07-27 #338 @arfon
@mwmclean - I'm having some major problems compiling the `paper.md` and `paper.bib` file you submitted, looks like it's failing on some kind of encoding issue:

```
/usr/local/var/rbenv/versions/2.4.1/lib/ruby/gems/2.4.0/gems/bibtex-ruby-4.4.4/lib/bibtex/lexer.rb:216:in `scan_until': invalid byte sequence in UTF-8 (ArgumentError)
	from /usr/local/var/rbenv/versions/2.4.1/lib/ruby/gems/2.4.0/gems/bibtex-ruby-4.4.4/lib/bibtex/lexer.rb:216:in `parse_meta'
	from /usr/local/var/rbenv/versions/2.4.1/lib/ruby/gems/2.4.0/gems/bibtex-ruby-4.4.4/lib/bibtex/lexer.rb:175:in `analyse'
	from bibtex.y:110:in `parse'
```

Can I just check whether you're doing anything funky with file encodings here?
