### 2017-07-27 #338 @mwmclean
Not that I know of.  Just looked at the `paper.md` and `paper.bib` again; can't see what the issue would be.
