### 2017-07-28 44f0567
change URLs for rOpenSci
