### 2017-08-01 bd92bf7
update coveralls badge
