### 2017-08-01 1d881a0
switch to github covr and codecov
