### 2017-08-03 5617b5b
back to CRAN covr
