### 2017-08-03 717ced2
fix covr install
